backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openai
groq
langchain-groq
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic
pydantic-settings
httpx
aiofiles
python-dateutil
pytest
pytest-asyncio
